# yiyu-server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "一隅 yiyu-server API",
    "description": "「一隅」后端 API。统一响应包装 envelope:{success, code, message, data, timestamp};成功 code=0,失败 code=HTTP 状态码,message 为中文人话。除 auth/login、auth/register 外全部要求 Bearer JWT;admin 组另要求 role=admin。时间格式 YYYY-MM-DD HH:mm(Asia/Shanghai),日期 YYYY-MM-DD,金额 number(元,≤2 位小数)。限流:全局 60 次/分,登录/注册 5 次/分(429)。",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "本地开发" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/ledger/books": {
      "get": {
        "tags": ["ledger"],
        "summary": "账本列表",
        "description": "monthExpense 为当月支出实时聚合;默认账本在前。",
        "responses": {
          "200": {
            "description": "账本列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Book" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ledger"],
        "summary": "新建账本",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "icon"],
                "properties": {
                  "name": { "type": "string" },
                  "icon": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新建的账本",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Book" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ledger/categories": {
      "get": {
        "tags": ["ledger"],
        "summary": "分类树",
        "description": "预置在前,custom=true 为我的自定义;带 type 返回单套树,省略 type 时一次返回两套 {expense:[], income:[]}。",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["expense", "income"] }
          }
        ],
        "responses": {
          "200": {
            "description": "根分类树",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": { "$ref": "#/components/schemas/Category" }
                            },
                            {
                              "type": "object",
                              "properties": {
                                "expense": {
                                  "type": "array",
                                  "items": { "$ref": "#/components/schemas/Category" }
                                },
                                "income": {
                                  "type": "array",
                                  "items": { "$ref": "#/components/schemas/Category" }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ledger"],
        "summary": "新增自定义分类",
        "description": "同 type 重名(含预置)→ 400「该分类名已存在」。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "name", "icon"],
                "properties": {
                  "type": { "type": "string", "enum": ["expense", "income"] },
                  "name": { "type": "string" },
                  "icon": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新建的分类",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Category" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ledger/categories/{id}": {
      "delete": {
        "tags": ["ledger"],
        "summary": "删除自定义分类",
        "description": "仅 custom 可删(他人/预置 403);删除后其流水 categoryId 置 null,展示靠 categoryName 兜底。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Empty" }
        }
      }
    },
    "/ledger/transactions": {
      "get": {
        "tags": ["ledger"],
        "summary": "流水列表(服务端分页 + 筛选)",
        "description": "bookId 必填(他人账本 403),to 含当日,时间倒序;type/categoryId/keyword 为服务端筛选(keyword 模糊匹配备注)。",
        "parameters": [
          { "name": "bookId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "from", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "type", "in": "query", "required": false, "schema": { "type": "string", "enum": ["expense", "income"] } },
          { "name": "categoryId", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "keyword", "in": "query", "required": false, "schema": { "type": "string", "maxLength": 60 } },
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "pageSize", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "分页流水",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/PagedTransactions" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ledger"],
        "summary": "记一笔",
        "description": "categoryName 由服务端派生;type 与分类不匹配 400;金额须 >0 且 ≤2 位小数。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "amount", "categoryId", "date", "bookId"],
                "properties": {
                  "type": { "type": "string", "enum": ["expense", "income"] },
                  "amount": { "type": "number", "exclusiveMinimum": 0 },
                  "categoryId": { "type": "string" },
                  "date": { "type": "string", "format": "date" },
                  "note": { "type": "string", "nullable": true },
                  "bookId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新建的流水",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Transaction" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ledger/transactions/{id}": {
      "patch": {
        "tags": ["ledger"],
        "summary": "更新流水",
        "description": "部分更新;非本人 403。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["expense", "income"] },
                  "amount": { "type": "number", "exclusiveMinimum": 0 },
                  "categoryId": { "type": "string" },
                  "date": { "type": "string", "format": "date" },
                  "note": { "type": "string", "nullable": true },
                  "bookId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的流水",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Transaction" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["ledger"],
        "summary": "删除流水",
        "description": "非本人 403。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Empty" }
        }
      }
    },
    "/ledger/reports": {
      "get": {
        "tags": ["ledger"],
        "summary": "报表聚合",
        "description": "三参均必填;from>to 或区间>10 年 → 400;他人账本 403。子分类归并到根分类名,金额倒序;daily 仅区间 ≤92 天时返回。",
        "parameters": [
          { "name": "bookId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "from", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "报表聚合结果",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Reports" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["admin"],
        "summary": "用户列表",
        "description": "keyword 模糊匹配用户名/昵称(不分大小写)。",
        "parameters": [
          { "name": "keyword", "in": "query", "required": false, "schema": { "type": "string" } },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["admin", "user"] }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["active", "disabled"] }
          }
        ],
        "responses": {
          "200": {
            "description": "用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/AdminUser" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "直接建号",
        "description": "不消耗邀请码,记 security 日志;用户名冲突 409。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "name", "avatar", "role", "password"],
                "properties": {
                  "username": { "type": "string", "pattern": "^[a-zA-Z0-9_]{2,20}$" },
                  "name": { "type": "string" },
                  "avatar": { "type": "string" },
                  "role": { "type": "string", "enum": ["admin", "user"] },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新建的用户",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AdminUser" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "tags": ["admin"],
        "summary": "改角色 / 停用启用",
        "description": "操作自己 400「不能修改自己的角色或状态」;停用即时生效(该用户现有 token 因 strategy 查库而 401)。",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": { "type": "string", "enum": ["admin", "user"] },
                  "status": { "type": "string", "enum": ["active", "disabled"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的用户",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AdminUser" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/logs": {
      "get": {
        "tags": ["admin"],
        "summary": "系统日志(服务端分页)",
        "description": "时间倒序,pageSize ≤100。",
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "pageSize", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
          { "name": "module", "in": "query", "required": false, "schema": { "type": "string", "enum": ["auth", "ledger", "profile", "admin"] } },
          { "name": "action", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "operatorId", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页日志",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Envelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/PagedLogs" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Empty": {
        "description": "成功(data 为 null)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Envelope" }
          }
        }
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "description": "统一响应包装;成功 code=0/message=''(空串),失败 code=HTTP 状态码/message 中文人话。",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "data": { "nullable": true },
          "timestamp": { "type": "integer", "description": "毫秒时间戳" }
        }
      },
...
      "Book": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "icon": { "type": "string" },
          "monthExpense": { "type": "number", "description": "当月支出实时聚合" },
          "isDefault": { "type": "boolean" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "icon": { "type": "string" },
          "custom": { "type": "boolean", "nullable": true },
          "children": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" }
              }
            }
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "enum": ["expense", "income"] },
          "amount": { "type": "number" },
          "categoryId": { "type": "string", "description": "已删分类为 ''" },
          "categoryName": { "type": "string", "description": "服务端派生" },
          "bookId": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "note": { "type": "string", "nullable": true }
        }
      },
      "PagedTransactions": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Transaction" }
          }
        }
      },
      "Reports": {
        "type": "object",
        "properties": {
          "stats": {
            "type": "object",
            "properties": {
              "income": { "type": "number" },
              "expense": { "type": "number" },
              "balance": { "type": "number" },
              "count": { "type": "integer" }
            }
          },
          "daily": {
            "type": "array",
            "description": "仅区间 ≤92 天时返回",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string", "format": "date" },
                "income": { "type": "number" },
                "expense": { "type": "number" }
              }
            }
          },
          "monthly": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "month": { "type": "string", "description": "YYYY-MM" },
                "income": { "type": "number" },
                "expense": { "type": "number" }
              }
            }
          },
          "categories": {
            "type": "object",
            "properties": {
              "expense": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/CategorySlice" }
              },
              "income": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/CategorySlice" }
              }
            }
          }
        }
      },
      "CategorySlice": {
        "type": "object",
        "description": "子分类归并到根分类名,金额倒序",
        "properties": {
          "name": { "type": "string" },
          "value": { "type": "number" }
        }
      },
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "name": { "type": "string" },
          "avatar": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "user"] },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "registeredAt": { "type": "string", "description": "YYYY-MM-DD HH:mm" },
          "lastActiveAt": { "type": "string", "nullable": true, "description": "YYYY-MM-DD HH:mm" }
        }
      },
      "AdminLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "time": { "type": "string", "description": "YYYY-MM-DD HH:mm" },
          "moduleId": { "type": "string", "enum": ["auth", "ledger", "profile", "admin"] },
          "action": { "type": "string" },
          "operatorId": { "type": "string" },
          "summary": { "type": "string" },
          "operator": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "nickname": { "type": "string" },
              "avatar": { "type": "string" }
            }
          }
        }
      },
      "PagedLogs": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminLog" }
          }
        }
      },
...
    }
  }
}